@@ -40,23 +40,36 @@ public class Recipes extends RecipeProvider {
         add('G', Items.GLASS);
         add('I', Tags.Items.STORAGE_BLOCKS_IRON);
         add('W', Items.GRAY_WOOL);
-        add('O', Tags.Items.OBSIDIAN);
+        add('O', Tags.Items.OBSIDIANS);
         add('C', Items.COBBLESTONE);
-        add('A', ing("forge:ingots/aluminum"));
+        add('A', ing("c:ingots/aluminum"));
 
     }
 
 
-    protected static void buildShapeless(RecipeOutput consumer, RecipeCategory cat, ItemLike result, int count, Item criterion, Ingredient... ingredients) {
+    protected static void buildShapeless(RecipeOutput consumer,
+                                         RecipeCategory cat,
+                                         ItemLike result,
+                                         int count,
+                                         Item criterion,
+                                         Ingredient... ingredients) {
         buildShapeless(consumer, cat, result, count, criterion, "", ingredients);
     }
 
-    protected static void buildShapeless(RecipeOutput consumer, RecipeCategory cat, ItemLike result, int count, Item criterion, String save, Ingredient... ingredients) {
-        Allomancy.LOGGER.debug("Creating Shapeless Recipe for " + BuiltInRegistries.ITEM.getKey(result.asItem()) + " " + save);
+    protected static void buildShapeless(RecipeOutput consumer,
+                                         RecipeCategory cat,
+                                         ItemLike result,
+                                         int count,
+                                         Item criterion,
+                                         String save,
+                                         Ingredient... ingredients) {
+        Allomancy.LOGGER.debug(
+                "Creating Shapeless Recipe for " + BuiltInRegistries.ITEM.getKey(result.asItem()) + " " + save);
 
         ShapelessRecipeBuilder builder = ShapelessRecipeBuilder.shapeless(cat, result, count);
 
-        builder.unlockedBy("has_" + BuiltInRegistries.ITEM.getKey(criterion).getPath(), InventoryChangeTrigger.TriggerInstance.hasItems(criterion));
+        builder.unlockedBy("has_" + BuiltInRegistries.ITEM.getKey(criterion).getPath(),
+                           InventoryChangeTrigger.TriggerInstance.hasItems(criterion));
 
         for (Ingredient ingredient : ingredients) {
             builder.requires(ingredient);
@@ -69,16 +82,25 @@ public class Recipes extends RecipeProvider {
         }
     }
 
-    protected static void buildSmeltingAndBlasting(RecipeOutput consumer, ItemLike result, ItemLike ingredient, float xp) {
-        Allomancy.LOGGER.debug("Creating Smelting and Blasting Recipe for " + BuiltInRegistries.ITEM.getKey(result.asItem()));
+    protected static void buildSmeltingAndBlasting(RecipeOutput consumer,
+                                                   ItemLike result,
+                                                   ItemLike ingredient,
+                                                   float xp) {
+        Allomancy.LOGGER.debug(
+                "Creating Smelting and Blasting Recipe for " + BuiltInRegistries.ITEM.getKey(result.asItem()));
 
-        SimpleCookingRecipeBuilder smelt = SimpleCookingRecipeBuilder.smelting(ing(ingredient), RecipeCategory.MISC, result, xp, 200);
-        SimpleCookingRecipeBuilder blast = SimpleCookingRecipeBuilder.blasting(ing(ingredient), RecipeCategory.MISC, result, xp, 100);
+        SimpleCookingRecipeBuilder smelt =
+                SimpleCookingRecipeBuilder.smelting(ing(ingredient), RecipeCategory.MISC, result, xp, 200);
+        SimpleCookingRecipeBuilder blast =
+                SimpleCookingRecipeBuilder.blasting(ing(ingredient), RecipeCategory.MISC, result, xp, 100);
 
-        smelt.unlockedBy("has_" + BuiltInRegistries.ITEM.getKey(ingredient.asItem()).getPath(), InventoryChangeTrigger.TriggerInstance.hasItems(ingredient));
-        blast.unlockedBy("has_" + BuiltInRegistries.ITEM.getKey(ingredient.asItem()).getPath(), InventoryChangeTrigger.TriggerInstance.hasItems(ingredient));
+        smelt.unlockedBy("has_" + BuiltInRegistries.ITEM.getKey(ingredient.asItem()).getPath(),
+                         InventoryChangeTrigger.TriggerInstance.hasItems(ingredient));
+        blast.unlockedBy("has_" + BuiltInRegistries.ITEM.getKey(ingredient.asItem()).getPath(),
+                         InventoryChangeTrigger.TriggerInstance.hasItems(ingredient));
 
-        var name = BuiltInRegistries.ITEM.getKey(result.asItem()) + "_from_" + BuiltInRegistries.ITEM.getKey(ingredient.asItem()).getPath();
+        var name = BuiltInRegistries.ITEM.getKey(result.asItem()) + "_from_" +
+                   BuiltInRegistries.ITEM.getKey(ingredient.asItem()).getPath();
 
         smelt.save(consumer, name);
         blast.save(consumer, name + "_from_blasting");
@@ -94,7 +116,7 @@ public class Recipes extends RecipeProvider {
     }
 
     protected static Ingredient ing(String tag) {
-        return Ingredient.of(ItemTags.create(new ResourceLocation(tag)));
+        return Ingredient.of(ItemTags.create(ResourceLocation.parse(tag)));
     }
 
     protected static Ingredient ing(TagKey<Item> tag) {
@@ -133,24 +155,32 @@ public class Recipes extends RecipeProvider {
         // Basic Shaped Recipes
         buildShaped(consumer, RecipeCategory.REDSTONE, ExtrasSetup.IRON_LEVER.get(), Items.IRON_INGOT, "s", "I");
         buildShaped(consumer, RecipeCategory.REDSTONE, ExtrasSetup.IRON_BUTTON.get(), Items.IRON_INGOT, "i", "I");
-        buildShapeless(consumer, RecipeCategory.REDSTONE, ExtrasSetup.INVERTED_IRON_BUTTON.get(), 1, ExtrasSetup.IRON_BUTTON_ITEM.get(), "inverted_from_iron_button",
+        buildShapeless(consumer, RecipeCategory.REDSTONE, ExtrasSetup.INVERTED_IRON_BUTTON.get(), 1,
+                       ExtrasSetup.IRON_BUTTON_ITEM.get(), "inverted_from_iron_button",
                        ing(ExtrasSetup.IRON_BUTTON.get()));
-        buildShapeless(consumer, RecipeCategory.REDSTONE, ExtrasSetup.IRON_BUTTON.get(), 1, ExtrasSetup.INVERTED_IRON_BUTTON_ITEM.get(), "iron_button_from_inverted",
+        buildShapeless(consumer, RecipeCategory.REDSTONE, ExtrasSetup.IRON_BUTTON.get(), 1,
+                       ExtrasSetup.INVERTED_IRON_BUTTON_ITEM.get(), "iron_button_from_inverted",
                        ing(ExtrasSetup.INVERTED_IRON_BUTTON.get()));
 
 
-        buildShaped(consumer, RecipeCategory.FOOD, ConsumeSetup.ALLOMANTIC_GRINDER.get(), Items.IRON_INGOT, "ggg", "iii", "ggg");
+        buildShaped(consumer, RecipeCategory.FOOD, ConsumeSetup.ALLOMANTIC_GRINDER.get(), Items.IRON_INGOT, "ggg",
+                    "iii", "ggg");
         buildShaped(consumer, RecipeCategory.FOOD, ConsumeSetup.VIAL.get(), 4, Items.GLASS, " S ", "G G", " G ");
-        buildShaped(consumer, RecipeCategory.COMBAT, CombatSetup.MISTCLOAK.get(), ConsumeSetup.VIAL.get(), "W W", "WWW", "WWW");
-        buildShaped(consumer, RecipeCategory.COMBAT, CombatSetup.ALUMINUM_HELMET.get(), MaterialsSetup.INGOTS.get(Metal.ALUMINUM.getIndex()).get(), "AAA", "A A");
+        buildShaped(consumer, RecipeCategory.COMBAT, CombatSetup.MISTCLOAK.get(), ConsumeSetup.VIAL.get(), "W W",
+                    "WWW", "WWW");
+        buildShaped(consumer, RecipeCategory.COMBAT, CombatSetup.ALUMINUM_HELMET.get(),
+                    MaterialsSetup.INGOTS.get(Metal.ALUMINUM.getIndex()).get(), "AAA", "A A");
 
-        buildShaped(consumer, RecipeCategory.COMBAT, CombatSetup.OBSIDIAN_DAGGER.get(), CombatSetup.MISTCLOAK.get(), "  O", " O ", "s  ");
-        buildShaped(consumer, RecipeCategory.COMBAT, CombatSetup.KOLOSS_BLADE.get(), ConsumeSetup.LERASIUM_NUGGET.get(), "CC", "CC", "sC");
+        buildShaped(consumer, RecipeCategory.COMBAT, CombatSetup.OBSIDIAN_DAGGER.get(), CombatSetup.MISTCLOAK.get(),
+                    "  O", " O ", "s  ");
+        buildShaped(consumer, RecipeCategory.COMBAT, CombatSetup.KOLOSS_BLADE.get(),
+                    ConsumeSetup.LERASIUM_NUGGET.get(), "CC", "CC", "sC");
 
         // Ore Recipes
         // order must be the same as MaterialsSetup.ORE_METALS
-        int[] ore_metal_indexes = {Metal.ALUMINUM.getIndex(), Metal.CADMIUM.getIndex(), Metal.CHROMIUM.getIndex(), MaterialsSetup.LEAD, MaterialsSetup.SILVER, Metal.TIN.getIndex(),
-                                   Metal.ZINC.getIndex()};
+        int[] ore_metal_indexes =
+                {Metal.ALUMINUM.getIndex(), Metal.CADMIUM.getIndex(), Metal.CHROMIUM.getIndex(), MaterialsSetup.LEAD,
+                 MaterialsSetup.SILVER, Metal.TIN.getIndex(), Metal.ZINC.getIndex()};
         float[] ore_metal_xp = {0.6F, 0.7F, 0.7F, 0.4F, 1.0F, 0.6F, 0.6F};
         for (int i = 0; i < MaterialsSetup.ORE_METALS.length; i++) {
             var raw = MaterialsSetup.RAW_ORE_ITEMS.get(i).get();
@@ -163,16 +193,20 @@ public class Recipes extends RecipeProvider {
         }
 
 
-        buildShapeless(consumer, RecipeCategory.MISC, ConsumeSetup.LERASIUM_NUGGET.get(), 1, Items.NETHER_STAR, ing(Items.NETHER_STAR), ing(Tags.Items.STORAGE_BLOCKS_GOLD));
+        buildShapeless(consumer, RecipeCategory.MISC, ConsumeSetup.LERASIUM_NUGGET.get(), 1, Items.NETHER_STAR,
+                       ing(Items.NETHER_STAR), ing(Tags.Items.STORAGE_BLOCKS_GOLD));
 
         // Most metal based recipes
         for (int i = 0; i < MaterialsSetup.METAL_ITEM_LEN; i++) {
 
             // Grinder recipes
             Item flake = MaterialsSetup.FLAKES.get(i).get();
-            String flakeType = BuiltInRegistries.ITEM.getKey(flake).getPath().substring(0, BuiltInRegistries.ITEM.getKey(flake).getPath().indexOf('_'));
-            buildShapeless(consumer, RecipeCategory.MISC, flake, 2, ConsumeSetup.ALLOMANTIC_GRINDER.get(), ing(ConsumeSetup.ALLOMANTIC_GRINDER.get()),
-                           ing("forge:" + "ingots/" + flakeType));
+            String flakeType = BuiltInRegistries.ITEM
+                    .getKey(flake)
+                    .getPath()
+                    .substring(0, BuiltInRegistries.ITEM.getKey(flake).getPath().indexOf('_'));
+            buildShapeless(consumer, RecipeCategory.MISC, flake, 2, ConsumeSetup.ALLOMANTIC_GRINDER.get(),
+                           ing(ConsumeSetup.ALLOMANTIC_GRINDER.get()), ing("c:" + "ingots/" + flakeType));
 
             if (i < Metal.values().length && Metal.getMetal(i).isVanilla()) {
                 continue;
@@ -185,73 +219,96 @@ public class Recipes extends RecipeProvider {
 
             // building up
             buildShapeless(consumer, RecipeCategory.BUILDING_BLOCKS, block, 1, ingot, repeat(ing(ingot), 9));
-            buildShapeless(consumer, RecipeCategory.MISC, ingot, 1, nugget, "allomancy:" + BuiltInRegistries.ITEM.getKey(ingot).getPath() + "_from_nuggets",
+            buildShapeless(consumer, RecipeCategory.MISC, ingot, 1, nugget,
+                           "allomancy:" + BuiltInRegistries.ITEM.getKey(ingot).getPath() + "_from_nuggets",
                            repeat(ing(nugget), 9));
 
             // breaking down
-            buildShapeless(consumer, RecipeCategory.MISC, ingot, 9, block, "allomancy:" + BuiltInRegistries.ITEM.getKey(ingot).getPath() + "_from_block", ing(block));
+            buildShapeless(consumer, RecipeCategory.MISC, ingot, 9, block,
+                           "allomancy:" + BuiltInRegistries.ITEM.getKey(ingot).getPath() + "_from_block", ing(block));
             buildShapeless(consumer, RecipeCategory.MISC, nugget, 9, ingot, ing(ingot));
         }
 
 
         // pattern recipes
         for (Metal mt : Metal.values()) {
-            buildShapeless(consumer, RecipeCategory.MISC, ExtrasSetup.PATTERN_ITEMS.get(mt.getIndex()).get(), 1, MaterialsSetup.FLAKES.get(mt.getIndex()).get(), ing(Items.PAPER),
+            buildShapeless(consumer, RecipeCategory.MISC, ExtrasSetup.PATTERN_ITEMS.get(mt.getIndex()).get(), 1,
+                           MaterialsSetup.FLAKES.get(mt.getIndex()).get(), ing(Items.PAPER),
                            ing(MaterialsSetup.FLAKES.get(mt.getIndex()).get()));
         }
 
         // Mixing/Alloying Recipes
         // GRINDER
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.STEEL.getIndex()).get(), 2, ConsumeSetup.ALLOMANTIC_GRINDER.get(), mixing_save("steel"),
-                       ing(Items.COAL), ing(MaterialsSetup.FLAKES.get(Metal.IRON.getIndex()).get()));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.STEEL.getIndex()).get(), 2,
+                       ConsumeSetup.ALLOMANTIC_GRINDER.get(), mixing_save("steel"), ing(Items.COAL),
+                       ing(MaterialsSetup.FLAKES.get(Metal.IRON.getIndex()).get()));
 
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.PEWTER.getIndex()).get(), 3, ConsumeSetup.ALLOMANTIC_GRINDER.get(), mixing_save("pewter"),
-                       repeatWith(ing(MaterialsSetup.FLAKES.get(Metal.TIN.getIndex()).get()), 2, ing(MaterialsSetup.FLAKES.get(MaterialsSetup.LEAD).get())));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.PEWTER.getIndex()).get(), 3,
+                       ConsumeSetup.ALLOMANTIC_GRINDER.get(), mixing_save("pewter"),
+                       repeatWith(ing(MaterialsSetup.FLAKES.get(Metal.TIN.getIndex()).get()), 2,
+                                  ing(MaterialsSetup.FLAKES.get(MaterialsSetup.LEAD).get())));
 
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.BRASS.getIndex()).get(), 4, ConsumeSetup.ALLOMANTIC_GRINDER.get(), mixing_save("brass"),
-                       repeatWith(ing(MaterialsSetup.FLAKES.get(Metal.COPPER.getIndex()).get()), 3, ing(MaterialsSetup.FLAKES.get(Metal.ZINC.getIndex()).get())));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.BRASS.getIndex()).get(), 4,
+                       ConsumeSetup.ALLOMANTIC_GRINDER.get(), mixing_save("brass"),
+                       repeatWith(ing(MaterialsSetup.FLAKES.get(Metal.COPPER.getIndex()).get()), 3,
+                                  ing(MaterialsSetup.FLAKES.get(Metal.ZINC.getIndex()).get())));
 
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.BRONZE.getIndex()).get(), 4, ConsumeSetup.ALLOMANTIC_GRINDER.get(), mixing_save("bronze"),
-                       repeatWith(ing(MaterialsSetup.FLAKES.get(Metal.COPPER.getIndex()).get()), 3, ing(MaterialsSetup.FLAKES.get(Metal.TIN.getIndex()).get())));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.BRONZE.getIndex()).get(), 4,
+                       ConsumeSetup.ALLOMANTIC_GRINDER.get(), mixing_save("bronze"),
+                       repeatWith(ing(MaterialsSetup.FLAKES.get(Metal.COPPER.getIndex()).get()), 3,
+                                  ing(MaterialsSetup.FLAKES.get(Metal.TIN.getIndex()).get())));
 
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.DURALUMIN.getIndex()).get(), 4, ConsumeSetup.ALLOMANTIC_GRINDER.get(),
-                       mixing_save("duralumin"),
-                       repeatWith(ing(MaterialsSetup.FLAKES.get(Metal.ALUMINUM.getIndex()).get()), 3, ing(MaterialsSetup.FLAKES.get(Metal.COPPER.getIndex()).get())));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.DURALUMIN.getIndex()).get(), 4,
+                       ConsumeSetup.ALLOMANTIC_GRINDER.get(), mixing_save("duralumin"),
+                       repeatWith(ing(MaterialsSetup.FLAKES.get(Metal.ALUMINUM.getIndex()).get()), 3,
+                                  ing(MaterialsSetup.FLAKES.get(Metal.COPPER.getIndex()).get())));
 
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.ELECTRUM.getIndex()).get(), 2, ConsumeSetup.ALLOMANTIC_GRINDER.get(), mixing_save("electrum"),
-                       ing(MaterialsSetup.FLAKES.get(Metal.GOLD.getIndex()).get()), ing(MaterialsSetup.FLAKES.get(MaterialsSetup.SILVER).get()));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.ELECTRUM.getIndex()).get(), 2,
+                       ConsumeSetup.ALLOMANTIC_GRINDER.get(), mixing_save("electrum"),
+                       ing(MaterialsSetup.FLAKES.get(Metal.GOLD.getIndex()).get()),
+                       ing(MaterialsSetup.FLAKES.get(MaterialsSetup.SILVER).get()));
 
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.BENDALLOY.getIndex()).get(), 3, ConsumeSetup.ALLOMANTIC_GRINDER.get(),
-                       mixing_save("bendalloy"),
-                       repeatWith(ing(MaterialsSetup.FLAKES.get(Metal.CADMIUM.getIndex()).get()), 2, ing(MaterialsSetup.FLAKES.get(MaterialsSetup.LEAD).get())));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.BENDALLOY.getIndex()).get(), 3,
+                       ConsumeSetup.ALLOMANTIC_GRINDER.get(), mixing_save("bendalloy"),
+                       repeatWith(ing(MaterialsSetup.FLAKES.get(Metal.CADMIUM.getIndex()).get()), 2,
+                                  ing(MaterialsSetup.FLAKES.get(MaterialsSetup.LEAD).get())));
 
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.NICROSIL.getIndex()).get(), 4, ConsumeSetup.ALLOMANTIC_GRINDER.get(), mixing_save("nicrosil"),
-                       repeatWith(ing(MaterialsSetup.FLAKES.get(Metal.CHROMIUM.getIndex()).get()), 3, ing(MaterialsSetup.FLAKES.get(Metal.IRON.getIndex()).get())));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.FLAKES.get(Metal.NICROSIL.getIndex()).get(), 4,
+                       ConsumeSetup.ALLOMANTIC_GRINDER.get(), mixing_save("nicrosil"),
+                       repeatWith(ing(MaterialsSetup.FLAKES.get(Metal.CHROMIUM.getIndex()).get()), 3,
+                                  ing(MaterialsSetup.FLAKES.get(Metal.IRON.getIndex()).get())));
 
         // ALLOYS
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.STEEL.getIndex()).get(), 4, Items.COAL, alloy_save("steel"),
-                       repeatWith(ing(Items.IRON_INGOT), 3, ing(Items.COAL)));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.STEEL.getIndex()).get(), 4,
+                       Items.COAL, alloy_save("steel"), repeatWith(ing(Items.IRON_INGOT), 3, ing(Items.COAL)));
 
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.PEWTER.getIndex()).get(), 3, MaterialsSetup.INGOTS.get(Metal.TIN.getIndex()).get(),
-                       alloy_save("pewter"), repeatWith(ing("forge:ingots/tin"), 2, ing("forge:ingots/lead")));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.PEWTER.getIndex()).get(), 3,
+                       MaterialsSetup.INGOTS.get(Metal.TIN.getIndex()).get(), alloy_save("pewter"),
+                       repeatWith(ing("c:ingots/tin"), 2, ing("c:ingots/lead")));
 
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.BRASS.getIndex()).get(), 4, MaterialsSetup.INGOTS.get(Metal.ZINC.getIndex()).get(),
-                       alloy_save("brass"), repeatWith(ing(Items.COPPER_INGOT), 3, ing("forge:ingots/zinc")));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.BRASS.getIndex()).get(), 4,
+                       MaterialsSetup.INGOTS.get(Metal.ZINC.getIndex()).get(), alloy_save("brass"),
+                       repeatWith(ing(Items.COPPER_INGOT), 3, ing("c:ingots/zinc")));
 
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.BRONZE.getIndex()).get(), 4, MaterialsSetup.INGOTS.get(Metal.TIN.getIndex()).get(),
-                       alloy_save("bronze"), repeatWith(ing(Items.COPPER_INGOT), 3, ing("forge:ingots/tin")));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.BRONZE.getIndex()).get(), 4,
+                       MaterialsSetup.INGOTS.get(Metal.TIN.getIndex()).get(), alloy_save("bronze"),
+                       repeatWith(ing(Items.COPPER_INGOT), 3, ing("c:ingots/tin")));
 
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.DURALUMIN.getIndex()).get(), 4, MaterialsSetup.INGOTS.get(Metal.ALUMINUM.getIndex()).get(),
-                       alloy_save("duralumin"), repeatWith(ing("forge:ingots/aluminum"), 3, ing(Items.COPPER_INGOT)));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.DURALUMIN.getIndex()).get(), 4,
+                       MaterialsSetup.INGOTS.get(Metal.ALUMINUM.getIndex()).get(), alloy_save("duralumin"),
+                       repeatWith(ing("c:ingots/aluminum"), 3, ing(Items.COPPER_INGOT)));
 
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.ELECTRUM.getIndex()).get(), 2, MaterialsSetup.INGOTS.get(MaterialsSetup.SILVER).get(),
-                       alloy_save("electrum"), ing("forge:ingots/silver"), ing(Items.GOLD_INGOT));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.ELECTRUM.getIndex()).get(), 2,
+                       MaterialsSetup.INGOTS.get(MaterialsSetup.SILVER).get(), alloy_save("electrum"),
+                       ing("c:ingots/silver"), ing(Items.GOLD_INGOT));
 
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.BENDALLOY.getIndex()).get(), 3, MaterialsSetup.INGOTS.get(Metal.CADMIUM.getIndex()).get(),
-                       alloy_save("bendalloy"), repeatWith(ing("forge:ingots/cadmium"), 2, ing("forge:ingots/lead")));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.BENDALLOY.getIndex()).get(), 3,
+                       MaterialsSetup.INGOTS.get(Metal.CADMIUM.getIndex()).get(), alloy_save("bendalloy"),
+                       repeatWith(ing("c:ingots/cadmium"), 2, ing("c:ingots/lead")));
 
-        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.NICROSIL.getIndex()).get(), 4, MaterialsSetup.INGOTS.get(Metal.CHROMIUM.getIndex()).get(),
-                       alloy_save("nicrosil"), repeatWith(ing("forge:ingots/chromium"), 3, ing(Items.IRON_INGOT)));
+        buildShapeless(consumer, RecipeCategory.MISC, MaterialsSetup.INGOTS.get(Metal.NICROSIL.getIndex()).get(), 4,
+                       MaterialsSetup.INGOTS.get(Metal.CHROMIUM.getIndex()).get(), alloy_save("nicrosil"),
+                       repeatWith(ing("c:ingots/chromium"), 3, ing(Items.IRON_INGOT)));
 
 
         Allomancy.LOGGER.debug("Creating Shaped Recipe for allomancy:coin_bag");
@@ -272,12 +329,18 @@ public class Recipes extends RecipeProvider {
 
     }
 
-    protected void buildShaped(RecipeOutput consumer, RecipeCategory cat, ItemLike result, int count, Item criterion, String... lines) {
+    protected void buildShaped(RecipeOutput consumer,
+                               RecipeCategory cat,
+                               ItemLike result,
+                               int count,
+                               Item criterion,
+                               String... lines) {
         Allomancy.LOGGER.debug("Creating Shaped Recipe for " + BuiltInRegistries.ITEM.getKey(result.asItem()));
 
         ShapedRecipeBuilder builder = ShapedRecipeBuilder.shaped(cat, result, count);
 
-        builder.unlockedBy("has_" + BuiltInRegistries.ITEM.getKey(criterion).getPath(), InventoryChangeTrigger.TriggerInstance.hasItems(criterion));
+        builder.unlockedBy("has_" + BuiltInRegistries.ITEM.getKey(criterion).getPath(),
+                           InventoryChangeTrigger.TriggerInstance.hasItems(criterion));
         builder.showNotification(true);
 
         Set<Character> characters = new HashSet<>();
@@ -295,7 +358,11 @@ public class Recipes extends RecipeProvider {
         builder.save(consumer);
     }
 
-    protected void buildShaped(RecipeOutput consumer, RecipeCategory cat, ItemLike result, Item criterion, String... lines) {
+    protected void buildShaped(RecipeOutput consumer,
+                               RecipeCategory cat,
+                               ItemLike result,
+                               Item criterion,
+                               String... lines) {
         buildShaped(consumer, cat, result, 1, criterion, lines);
     }
 
