@@ -1,12 +1,16 @@
 package com.legobmw99.allomancy.modules.consumables.item;
 
+import com.legobmw99.allomancy.Allomancy;
 import com.legobmw99.allomancy.api.enums.Metal;
 import com.legobmw99.allomancy.modules.consumables.ConsumeSetup;
+import com.legobmw99.allomancy.modules.consumables.item.component.FlakeStorage;
 import com.legobmw99.allomancy.modules.powers.data.AllomancerAttachment;
 import com.legobmw99.allomancy.modules.powers.data.AllomancerData;
 import com.legobmw99.allomancy.util.ItemDisplay;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.core.component.DataComponents;
+import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.world.InteractionHand;
@@ -16,42 +20,49 @@ import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.*;
+import net.minecraft.world.item.component.CustomData;
+import net.minecraft.world.item.component.CustomModelData;
 import net.minecraft.world.level.Level;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.api.distmarker.OnlyIn;
 
-import javax.annotation.Nullable;
 import java.util.List;
+
+import static com.legobmw99.allomancy.modules.consumables.ConsumeSetup.FLAKE_STORAGE;
 
 public class VialItem extends Item {
 
+    public static final CustomModelData FILLED_MODEL_DATA = new CustomModelData(1);
+
     public VialItem() {
-        super(new Item.Properties().stacksTo(32));
+        super(new Item.Properties().stacksTo(32).rarity(Rarity.COMMON));
     }
 
 
     @Override
     public ItemStack finishUsingItem(ItemStack stack, Level world, LivingEntity livingEntity) {
+        FlakeStorage storage = stack.get(FLAKE_STORAGE);
 
-        if (!stack.hasTag() || !livingEntity.hasData(AllomancerAttachment.ALLOMANCY_DATA)) {
+        if (storage == null || !livingEntity.hasData(AllomancerAttachment.ALLOMANCY_DATA)) {
             return stack;
         }
 
 
         var data = livingEntity.getData(AllomancerAttachment.ALLOMANCY_DATA);
         for (Metal mt : Metal.values()) {
-            if (stack.getTag().contains(mt.getName()) && stack.getTag().getBoolean(mt.getName())) {
-                if (data.getStored(mt) < AllomancerData.MAX_STORAGE) {
-                    data.incrementStored(mt);
-                }
+            if (storage.contains(mt) && data.getStored(mt) < AllomancerData.MAX_STORAGE) {
+                data.incrementStored(mt);
             }
         }
+
 
         if (!((Player) (livingEntity)).getAbilities().instabuild) {
             stack.shrink(1);
 
             if (!((Player) livingEntity).getInventory().add(new ItemStack(ConsumeSetup.VIAL.get(), 1))) {
-                world.addFreshEntity(new ItemEntity(world, livingEntity.getX(), livingEntity.getY(), livingEntity.getZ(), new ItemStack(ConsumeSetup.VIAL.get(), 1)));
+                world.addFreshEntity(
+                        new ItemEntity(world, livingEntity.getX(), livingEntity.getY(), livingEntity.getZ(),
+                                       new ItemStack(ConsumeSetup.VIAL.get(), 1)));
             }
         }
 
@@ -59,9 +70,10 @@ public class VialItem extends Item {
     }
 
     @Override
-    public int getUseDuration(ItemStack stack) {
+    public int getUseDuration(ItemStack stack, LivingEntity livingEntity) {
         return 6;
     }
+
 
     @Override
     public UseAnim getUseAnimation(ItemStack stack) {
@@ -76,9 +88,10 @@ public class VialItem extends Item {
         var data = playerIn.getData(AllomancerAttachment.ALLOMANCY_DATA);
         //If all the ones being filled are full, don't allow
         int filling = 0, full = 0;
-        if (itemStackIn.hasTag()) {
+        FlakeStorage storage = itemStackIn.get(FLAKE_STORAGE);
+        if (storage != null) {
             for (Metal mt : Metal.values()) {
-                if (itemStackIn.getTag().contains(mt.getName()) && itemStackIn.getTag().getBoolean(mt.getName())) {
+                if (storage.contains(mt)) {
                     filling++;
                     if (data.getStored(mt) >= AllomancerData.MAX_STORAGE) {
                         full++;
@@ -96,33 +109,77 @@ public class VialItem extends Item {
 
     @OnlyIn(Dist.CLIENT)
     @Override
-    public void appendHoverText(ItemStack stack, @Nullable Level worldIn, List<Component> tooltip, TooltipFlag flagIn) {
-        super.appendHoverText(stack, worldIn, tooltip, flagIn);
-        if (stack.hasTag()) {
+    public void appendHoverText(ItemStack stack,
+                                Item.TooltipContext ctx,
+                                List<Component> tooltip,
+                                TooltipFlag flagIn) {
+        super.appendHoverText(stack, ctx, tooltip, flagIn);
+
+        FlakeStorage storage = stack.get(FLAKE_STORAGE);
+        if (storage != null) {
             boolean full_display = Screen.hasShiftDown();
             int count = 0;
             for (Metal mt : Metal.values()) {
-                if (stack.getTag().getBoolean(mt.getName())) {
+
+
+                if (storage.contains(mt)) {
                     count++;
                     if (full_display) {
-                        MutableComponent metal = ItemDisplay.addColorToText("metals." + mt.getName(), ChatFormatting.GRAY);
+                        MutableComponent metal =
+                                ItemDisplay.addColorToText("metals." + mt.getName(), ChatFormatting.GRAY);
                         tooltip.add(metal);
                     }
                 }
             }
             if (!full_display) {
-                MutableComponent lcount = ItemDisplay.addColorToText("item.allomancy.vial.lore_count", ChatFormatting.GRAY, count);
+                MutableComponent lcount =
+                        ItemDisplay.addColorToText("item.allomancy.vial.lore_count", ChatFormatting.GRAY, count);
                 tooltip.add(lcount);
-                MutableComponent linst = ItemDisplay.addColorToText("item.allomancy.vial.lore_inst", ChatFormatting.GRAY);
+                MutableComponent linst =
+                        ItemDisplay.addColorToText("item.allomancy.vial.lore_inst", ChatFormatting.GRAY);
                 tooltip.add(linst);
 
             }
         }
     }
 
-    @Override
-    public Rarity getRarity(ItemStack stack) {
-        return stack.hasTag() ? Rarity.UNCOMMON : Rarity.COMMON;
+    public static void fillVial(ItemStack stack, FlakeStorage storage) {
+        stack.set(FLAKE_STORAGE, storage);
+        if (storage == null) {
+            stack.set(DataComponents.CUSTOM_MODEL_DATA, null);
+            stack.set(DataComponents.RARITY, Rarity.COMMON);
+        } else {
+            stack.set(DataComponents.CUSTOM_MODEL_DATA, FILLED_MODEL_DATA);
+            stack.set(DataComponents.RARITY, Rarity.UNCOMMON);
+        }
+
     }
 
+    /**
+     * TEMPORARY: Used to port pre-1.20.5 worlds to post.
+     * Loads custom NBT data and convets it to the data component.
+     */
+    @Override
+    public void verifyComponentsAfterLoad(ItemStack pStack) {
+        super.verifyComponentsAfterLoad(pStack);
+        if (pStack.has(FLAKE_STORAGE)) {
+            return;
+        }
+
+        CustomData customData = pStack.get(DataComponents.CUSTOM_DATA);
+        if (customData != null && customData.contains("steel")) {
+
+            CompoundTag tag = customData.copyTag();
+            Allomancy.LOGGER.info("Found old custom item data for vial: {}", tag.toString());
+            FlakeStorage.Mutable storage = new FlakeStorage.Mutable();
+            for (Metal mt : Metal.values()) {
+                if (tag.getBoolean(mt.getName())) {
+                    storage.add(mt);
+                }
+                tag.remove(mt.getName());
+            }
+            fillVial(pStack, storage.toImmutable());
+            pStack.set(DataComponents.CUSTOM_DATA, CustomData.of(tag));
+        }
+    }
 }
