@@ -9,7 +9,6 @@ import net.minecraft.advancements.critereon.ContextAwarePredicate;
 import net.minecraft.advancements.critereon.EntityPredicate;
 import net.minecraft.advancements.critereon.SimpleCriterionTrigger;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.util.ExtraCodecs;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.level.storage.loot.LootContext;
 
@@ -30,20 +29,33 @@ public class MetalUsedOnEntityTrigger extends SimpleCriterionTrigger<MetalUsedOn
         this.trigger(player, p_48112_ -> p_48112_.matches(lootcontext, mt, enhanced));
     }
 
-    public record TriggerInstance(Optional<ContextAwarePredicate> player, Optional<ContextAwarePredicate> entityPredicate, Metal mt,
+    public record TriggerInstance(Optional<ContextAwarePredicate> player,
+                                  Optional<ContextAwarePredicate> entityPredicate, Metal mt,
                                   Optional<Boolean> enhanced) implements SimpleCriterionTrigger.SimpleInstance {
         public static final Codec<TriggerInstance> CODEC = RecordCodecBuilder.create(builder -> builder
-                .group(ExtraCodecs.strictOptionalField(EntityPredicate.ADVANCEMENT_CODEC, "player").forGetter(TriggerInstance::player),
-                       ExtraCodecs.strictOptionalField(EntityPredicate.ADVANCEMENT_CODEC, "entity").forGetter(TriggerInstance::entityPredicate),
-                       Metal.CODEC.fieldOf("metal").forGetter(TriggerInstance::mt), ExtraCodecs.strictOptionalField(Codec.BOOL, "enhanced").forGetter(TriggerInstance::enhanced))
+                .group(EntityPredicate.ADVANCEMENT_CODEC.optionalFieldOf("player").forGetter(TriggerInstance::player),
+                       EntityPredicate.ADVANCEMENT_CODEC
+                               .optionalFieldOf("entity")
+                               .forGetter(TriggerInstance::entityPredicate),
+                       Metal.CODEC.fieldOf("metal").forGetter(TriggerInstance::mt),
+                       Codec.BOOL.optionalFieldOf("enhanced").forGetter(TriggerInstance::enhanced))
                 .apply(builder, TriggerInstance::new));
 
-        public static Criterion<TriggerInstance> instance(Optional<ContextAwarePredicate> player, Optional<ContextAwarePredicate> entityPredicate, Metal mt) {
-            return ExtrasSetup.METAL_USED_ON_ENTITY_TRIGGER.get().createCriterion(new TriggerInstance(player, entityPredicate, mt, Optional.empty()));
+        public static Criterion<TriggerInstance> instance(Optional<ContextAwarePredicate> player,
+                                                          Optional<ContextAwarePredicate> entityPredicate,
+                                                          Metal mt) {
+            return ExtrasSetup.METAL_USED_ON_ENTITY_TRIGGER
+                    .get()
+                    .createCriterion(new TriggerInstance(player, entityPredicate, mt, Optional.empty()));
         }
 
-        public static Criterion<TriggerInstance> instance(Optional<ContextAwarePredicate> player, Optional<ContextAwarePredicate> entityPredicate, Metal mt, boolean enhanced) {
-            return ExtrasSetup.METAL_USED_ON_ENTITY_TRIGGER.get().createCriterion(new TriggerInstance(player, entityPredicate, mt, Optional.of(enhanced)));
+        public static Criterion<TriggerInstance> instance(Optional<ContextAwarePredicate> player,
+                                                          Optional<ContextAwarePredicate> entityPredicate,
+                                                          Metal mt,
+                                                          boolean enhanced) {
+            return ExtrasSetup.METAL_USED_ON_ENTITY_TRIGGER
+                    .get()
+                    .createCriterion(new TriggerInstance(player, entityPredicate, mt, Optional.of(enhanced)));
         }
 
         boolean matches(LootContext entity, Metal mt, boolean enhanced) {
