@@ -9,12 +9,12 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
-import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.Explosion;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.LeverBlock;
@@ -22,8 +22,8 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.gameevent.GameEvent;
 import net.minecraft.world.phys.BlockHitResult;
 
-import javax.annotation.Nullable;
 import java.util.List;
+import java.util.function.BiConsumer;
 
 public class IronLeverBlock extends LeverBlock implements IAllomanticallyUsableBlock {
 
@@ -40,10 +40,11 @@ public class IronLeverBlock extends LeverBlock implements IAllomanticallyUsableB
             return true;
         }
         if (isPush == state.getValue(POWERED)) {
-            BlockState blockstate = this.pull(state, level, pos);
-            float f = blockstate.getValue(POWERED) ? 0.6F : 0.5F;
+            this.pull(state, level, pos, player);
+            float f = state.getValue(POWERED) ? 0.6F : 0.5F;
             level.playSound(null, pos, SoundEvents.LEVER_CLICK, SoundSource.BLOCKS, 0.3F, f);
-            level.gameEvent(player, blockstate.getValue(POWERED) ? GameEvent.BLOCK_ACTIVATE : GameEvent.BLOCK_DEACTIVATE, pos);
+            level.gameEvent(player, state.getValue(POWERED) ? GameEvent.BLOCK_ACTIVATE : GameEvent.BLOCK_DEACTIVATE,
+                            pos);
             return true;
 
         }
@@ -51,14 +52,29 @@ public class IronLeverBlock extends LeverBlock implements IAllomanticallyUsableB
     }
 
     @Override
-    public InteractionResult use(BlockState state, Level worldIn, BlockPos pos, Player player, InteractionHand handIn, BlockHitResult hit) {
+    protected InteractionResult useWithoutItem(BlockState pState,
+                                               Level pLevel,
+                                               BlockPos pPos,
+                                               Player pPlayer,
+                                               BlockHitResult pHitResult) {
         return InteractionResult.FAIL;
     }
 
+    @Override
+    protected void onExplosionHit(BlockState pState,
+                                  Level pLevel,
+                                  BlockPos pPos,
+                                  Explosion pExplosion,
+                                  BiConsumer<ItemStack, BlockPos> pDropConsumer) {
+
+    }
 
     @Override
-    public void appendHoverText(ItemStack stack, @Nullable BlockGetter worldIn, List<Component> tooltip, TooltipFlag flagIn) {
-        super.appendHoverText(stack, worldIn, tooltip, flagIn);
+    public void appendHoverText(ItemStack stack,
+                                Item.TooltipContext ctx,
+                                List<Component> tooltip,
+                                TooltipFlag flagIn) {
+        super.appendHoverText(stack, ctx, tooltip, flagIn);
         Component lore = ItemDisplay.addColorToText("block.allomancy.iron_activation.lore", ChatFormatting.GRAY);
         tooltip.add(lore);
     }
