@@ -3,10 +3,12 @@ package com.legobmw99.allomancy.modules.materials.world;
 import com.google.common.base.Suppliers;
 import com.legobmw99.allomancy.modules.combat.CombatSetup;
 import com.mojang.serialization.Codec;
+import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
-import net.minecraft.nbt.CompoundTag;
+import net.minecraft.core.component.DataComponents;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.component.Unbreakable;
 import net.minecraft.world.level.storage.loot.LootContext;
 import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
 import net.neoforged.neoforge.common.loot.IGlobalLootModifier;
@@ -18,8 +20,10 @@ import java.util.function.Supplier;
 public class DaggerLootModifier extends LootModifier {
 
     public final int chance_one_in;
-    public static final Supplier<Codec<DaggerLootModifier>> CODEC = Suppliers.memoize(
-            () -> RecordCodecBuilder.create(inst -> codecStart(inst).and(Codec.INT.fieldOf("chance_one_in").forGetter(t -> t.chance_one_in)).apply(inst, DaggerLootModifier::new)));
+    public static final Supplier<MapCodec<DaggerLootModifier>> CODEC = Suppliers.memoize(
+            () -> RecordCodecBuilder.mapCodec(inst -> codecStart(inst)
+                    .and(Codec.INT.fieldOf("chance_one_in").forGetter(t -> t.chance_one_in))
+                    .apply(inst, DaggerLootModifier::new)));
 
     public DaggerLootModifier(LootItemCondition[] conditionsIn, int chance_one_in) {
         super(conditionsIn);
@@ -27,19 +31,18 @@ public class DaggerLootModifier extends LootModifier {
     }
 
     @Override
-    protected @NotNull ObjectArrayList<ItemStack> doApply(ObjectArrayList<ItemStack> generatedLoot, LootContext context) {
+    protected @NotNull ObjectArrayList<ItemStack> doApply(ObjectArrayList<ItemStack> generatedLoot,
+                                                          LootContext context) {
         if (context.getRandom().nextInt(this.chance_one_in) == 0) {
             ItemStack dagger = new ItemStack(CombatSetup.OBSIDIAN_DAGGER.get());
-            CompoundTag nbt = new CompoundTag();
-            nbt.putBoolean("Unbreakable", true);
-            dagger.setTag(nbt);
+            dagger.set(DataComponents.UNBREAKABLE, new Unbreakable(true));
             generatedLoot.add(dagger);
         }
         return generatedLoot;
     }
 
     @Override
-    public Codec<? extends IGlobalLootModifier> codec() {
+    public MapCodec<? extends IGlobalLootModifier> codec() {
         return CODEC.get();
     }
 }
