@@ -3,6 +3,7 @@ package com.legobmw99.allomancy.modules.materials.world;
 import com.google.common.base.Suppliers;
 import com.legobmw99.allomancy.modules.consumables.ConsumeSetup;
 import com.mojang.serialization.Codec;
+import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import net.minecraft.world.item.ItemStack;
@@ -17,8 +18,10 @@ import java.util.function.Supplier;
 public class LerasiumLootModifier extends LootModifier {
 
     public final int chance_one_in;
-    public static final Supplier<Codec<LerasiumLootModifier>> CODEC = Suppliers.memoize(() -> RecordCodecBuilder.create(
-            inst -> codecStart(inst).and(Codec.INT.fieldOf("chance_one_in").forGetter(t -> t.chance_one_in)).apply(inst, LerasiumLootModifier::new)));
+    public static final Supplier<MapCodec<LerasiumLootModifier>> CODEC = Suppliers.memoize(
+            () -> RecordCodecBuilder.mapCodec(inst -> codecStart(inst)
+                    .and(Codec.INT.fieldOf("chance_one_in").forGetter(t -> t.chance_one_in))
+                    .apply(inst, LerasiumLootModifier::new)));
 
     public LerasiumLootModifier(LootItemCondition[] conditionsIn, int chance_one_in) {
         super(conditionsIn);
@@ -27,7 +30,8 @@ public class LerasiumLootModifier extends LootModifier {
 
 
     @Override
-    protected @NotNull ObjectArrayList<ItemStack> doApply(ObjectArrayList<ItemStack> generatedLoot, LootContext context) {
+    protected @NotNull ObjectArrayList<ItemStack> doApply(ObjectArrayList<ItemStack> generatedLoot,
+                                                          LootContext context) {
         if (context.getRandom().nextInt(this.chance_one_in) == 0) {
             generatedLoot.add(new ItemStack(ConsumeSetup.LERASIUM_NUGGET.get()));
         }
@@ -35,7 +39,7 @@ public class LerasiumLootModifier extends LootModifier {
     }
 
     @Override
-    public Codec<? extends IGlobalLootModifier> codec() {
+    public MapCodec<? extends IGlobalLootModifier> codec() {
         return CODEC.get();
     }
 }
