@@ -11,7 +11,6 @@ import com.legobmw99.allomancy.modules.powers.data.AllomancerAttachment;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.resources.sounds.SoundInstance;
-import net.minecraft.core.BlockPos;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.player.Player;
@@ -20,12 +19,8 @@ import net.minecraft.world.phys.Vec3;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.api.distmarker.OnlyIn;
 import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.neoforge.client.event.ComputeFovModifierEvent;
-import net.neoforged.neoforge.client.event.InputEvent;
-import net.neoforged.neoforge.client.event.MovementInputUpdateEvent;
-import net.neoforged.neoforge.client.event.RenderLevelStageEvent;
+import net.neoforged.neoforge.client.event.*;
 import net.neoforged.neoforge.client.event.sound.PlaySoundEvent;
-import net.neoforged.neoforge.event.TickEvent;
 import org.lwjgl.glfw.GLFW;
 
 public class ClientEventHandler {
@@ -33,10 +28,10 @@ public class ClientEventHandler {
 
     @OnlyIn(Dist.CLIENT)
     @SubscribeEvent
-    public static void onClientTick(final TickEvent.ClientTickEvent event) {
+    public static void onClientTick(final ClientTickEvent.Post event) {
         Minecraft mc = Minecraft.getInstance();
         // Run once per tick, only if in game, and only if there is a player
-        if (event.phase != TickEvent.Phase.END || mc.isPaused() || mc.player == null || !mc.player.isAlive()) {
+        if (mc.isPaused() || mc.player == null || !mc.player.isAlive()) {
             return;
         }
 
@@ -103,7 +98,7 @@ public class ClientEventHandler {
         if (data.isUninvested()) {
             return;
         }
-        float partialTick = event.getPartialTick();
+        float partialTick = event.getPartialTick().getGameTimeDeltaPartialTick(false);
 
         PoseStack stack = Rendering.prepareToDrawLines(event.getPoseStack(), partialTick);
 
@@ -111,7 +106,10 @@ public class ClientEventHandler {
         float theta = (float) ((player.getViewYRot(partialTick) + 90) * Math.PI / 180);
         float phi = Mth.clamp((float) ((player.getViewXRot(partialTick) + 90) * Math.PI / 180), 0.0001F, 3.14F);
 
-        Vec3 playervec = mc.cameraEntity.getEyePosition(partialTick).add(rho * Mth.sin(phi) * Mth.cos(theta), rho * Mth.cos(phi) - 0.35F, rho * Mth.sin(phi) * Mth.sin(theta));
+        Vec3 playervec = mc.cameraEntity
+                .getEyePosition(partialTick)
+                .add(rho * Mth.sin(phi) * Mth.cos(theta), rho * Mth.cos(phi) - 0.35F,
+                     rho * Mth.sin(phi) * Mth.sin(theta));
 
         /*********************************************
          * IRON AND STEEL LINES                      *
@@ -119,16 +117,21 @@ public class ClientEventHandler {
 
 
         if ((data.isBurning(Metal.IRON) || data.isBurning(Metal.STEEL))) {
-            tracking.forEachMetallicEntity(entity -> Rendering.drawMetalLine(stack, playervec, entity.position(), 1.5F, 0F, 0.6F, 1F));
+            tracking.forEachMetallicEntity(
+                    entity -> Rendering.drawMetalLine(stack, playervec, entity.position(), 1.5F, 0F, 0.6F, 1F));
 
-            tracking.forEachMetalBlob(blob -> Rendering.drawMetalLine(stack, playervec, blob.getCenter(), Mth.clamp(0.3F + blob.size() * 0.4F, 0.5F, 7.5F), 0F, 0.6F, 1F));
+            tracking.forEachMetalBlob(blob -> Rendering.drawMetalLine(stack, playervec, blob.getCenter(),
+                                                                      Mth.clamp(0.3F + blob.size() * 0.4F, 0.5F,
+                                                                                7.5F), 0F, 0.6F, 1F));
         }
 
         /*********************************************
          * BRONZE LINES                              *
          *********************************************/
         if ((data.isBurning(Metal.BRONZE) && (data.isEnhanced() || !data.isBurning(Metal.COPPER)))) {
-            tracking.forEachSeeked(playerEntity -> Rendering.drawMetalLine(stack, playervec, playerEntity.position(), 5.0F, 0.7F, 0.15F, 0.15F));
+            tracking.forEachSeeked(
+                    playerEntity -> Rendering.drawMetalLine(stack, playervec, playerEntity.position(), 5.0F, 0.7F,
+                                                            0.15F, 0.15F));
         }
 
         /*********************************************
@@ -137,17 +140,20 @@ public class ClientEventHandler {
         if (data.isBurning(Metal.GOLD)) {
             ResourceKey<Level> deathDim = data.getDeathDim();
             if (deathDim != null && player.level().dimension() == deathDim) {
-                Rendering.drawMetalLine(stack, playervec, Vec3.atCenterOf(data.getDeathLoc()), 3.0F, 0.9F, 0.85F, 0.0F);
+                Rendering.drawMetalLine(stack, playervec, Vec3.atCenterOf(data.getDeathLoc()), 3.0F, 0.9F, 0.85F,
+                                        0.0F);
             }
         }
         if (data.isBurning(Metal.ELECTRUM)) {
             ResourceKey<Level> spawnDim = data.getSpawnDim();
-            if (spawnDim == null && player.level().dimension() == Level.OVERWORLD) { // overworld, no spawn --> use world spawn
+            if (spawnDim == null &&
+                player.level().dimension() == Level.OVERWORLD) { // overworld, no spawn --> use world spawn
                 var levelData = player.level().getLevelData();
-                BlockPos spawnLoc = new BlockPos(levelData.getXSpawn(), levelData.getYSpawn(), levelData.getZSpawn());
-                Rendering.drawMetalLine(stack, playervec, Vec3.atCenterOf(spawnLoc), 3.0F, 0.7F, 0.8F, 0.2F);
+                Rendering.drawMetalLine(stack, playervec, Vec3.atCenterOf(levelData.getSpawnPos()), 3.0F, 0.7F, 0.8F,
+                                        0.2F);
             } else if (spawnDim != null && player.level().dimension() == spawnDim) {
-                Rendering.drawMetalLine(stack, playervec, Vec3.atCenterOf(data.getSpawnLoc()), 3.0F, 0.7F, 0.8F, 0.2F);
+                Rendering.drawMetalLine(stack, playervec, Vec3.atCenterOf(data.getSpawnLoc()), 3.0F, 0.7F, 0.8F,
+                                        0.2F);
             }
         }
 
