@@ -3,9 +3,11 @@ package com.legobmw99.allomancy.modules.powers.client;
 import com.legobmw99.allomancy.Allomancy;
 import com.legobmw99.allomancy.modules.powers.client.particle.SoundParticle;
 import com.legobmw99.allomancy.modules.powers.client.particle.SoundParticleData;
-import com.mojang.serialization.Codec;
+import com.mojang.serialization.MapCodec;
 import net.minecraft.core.particles.ParticleType;
 import net.minecraft.core.registries.Registries;
+import net.minecraft.network.RegistryFriendlyByteBuf;
+import net.minecraft.network.codec.StreamCodec;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
 import net.neoforged.neoforge.client.event.RegisterParticleProvidersEvent;
@@ -15,14 +17,22 @@ import net.neoforged.neoforge.registries.DeferredRegister;
 import java.util.function.Supplier;
 
 public class PowersClientSetup {
-    public static final DeferredRegister<ParticleType<?>> PARTICLES = DeferredRegister.create(Registries.PARTICLE_TYPE, Allomancy.MODID);
-    public static final Supplier<ParticleType<SoundParticleData>> SOUND_PARTICLE_TYPE = PARTICLES.register("sound_particle",
-                                                                                                           () -> new ParticleType<>(true, SoundParticleData.DESERIALIZER) {
-                                                                                                               @Override
-                                                                                                               public Codec<SoundParticleData> codec() {
-                                                                                                                   return SoundParticleData.CODEC;
-                                                                                                               }
-                                                                                                           });
+    public static final DeferredRegister<ParticleType<?>> PARTICLES =
+            DeferredRegister.create(Registries.PARTICLE_TYPE, Allomancy.MODID);
+    public static final Supplier<ParticleType<SoundParticleData>> SOUND_PARTICLE_TYPE =
+            PARTICLES.register("sound_particle", () -> new ParticleType<>(true) {
+
+
+                @Override
+                public MapCodec<SoundParticleData> codec() {
+                    return SoundParticleData.CODEC;
+                }
+
+                @Override
+                public StreamCodec<? super RegistryFriendlyByteBuf, SoundParticleData> streamCodec() {
+                    return SoundParticleData.STREAM_CODEC;
+                }
+            });
 
     public static void register(IEventBus bus) {
         PARTICLES.register(bus);
