@@ -5,22 +5,23 @@ import com.legobmw99.allomancy.api.enums.Metal;
 import com.legobmw99.allomancy.modules.powers.PowersConfig;
 import com.legobmw99.allomancy.modules.powers.data.AllomancerAttachment;
 import com.mojang.blaze3d.systems.RenderSystem;
+import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.LayeredDraw;
 import net.minecraft.client.gui.screens.ChatScreen;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.resources.ResourceLocation;
-import net.neoforged.neoforge.client.event.RegisterGuiOverlaysEvent;
-import net.neoforged.neoforge.client.gui.overlay.ExtendedGui;
-import net.neoforged.neoforge.client.gui.overlay.IGuiOverlay;
+import net.neoforged.neoforge.client.event.RegisterGuiLayersEvent;
 
 import java.awt.*;
 
-public class MetalOverlay implements IGuiOverlay {
+public class MetalOverlay implements LayeredDraw.Layer {
 
     private static final Point[] Frames = new Point[4];
-    private static final ResourceLocation meterLoc = new ResourceLocation("allomancy", "textures/gui/overlay/meter.png");
+    private static final ResourceLocation meterLoc =
+            ResourceLocation.fromNamespaceAndPath(Allomancy.MODID, "textures/gui/overlay/meter.png");
     private static int currentFrame = 0;
 
     static {
@@ -33,16 +34,23 @@ public class MetalOverlay implements IGuiOverlay {
 
     private MetalOverlay() {}
 
-    public static void registerGUI(final RegisterGuiOverlaysEvent evt) {
-        evt.registerAboveAll(new ResourceLocation(Allomancy.MODID, "metal_display"), new MetalOverlay());
+    public static void registerGUI(final RegisterGuiLayersEvent evt) {
+        evt.registerAboveAll(ResourceLocation.fromNamespaceAndPath(Allomancy.MODID, "metal_display"),
+                             new MetalOverlay());
     }
 
-    private static void blit(GuiGraphics graphics, int x, int y, float uOffset, float vOffset, int uWidth, int vHeight) {
+    private static void blit(GuiGraphics graphics,
+                             int x,
+                             int y,
+                             float uOffset,
+                             float vOffset,
+                             int uWidth,
+                             int vHeight) {
         graphics.blit(meterLoc, x, y, 0, uOffset, vOffset, uWidth, vHeight, 128, 128);
     }
 
     @Override
-    public void render(ExtendedGui gui, GuiGraphics guiGraphics, float partialTick, int screenWidth, int screenHeight) {
+    public void render(GuiGraphics gui, DeltaTracker deltaTracker) {
         Minecraft mc = Minecraft.getInstance();
         LocalPlayer player = mc.player;
 
@@ -63,8 +71,8 @@ public class MetalOverlay implements IGuiOverlay {
         }
 
 
-        int renderX = PowersConfig.overlay_position.get().getX(screenWidth);
-        int renderY = PowersConfig.overlay_position.get().getY(screenHeight);
+        int renderX = PowersConfig.overlay_position.get().getX(gui.guiWidth());
+        int renderY = PowersConfig.overlay_position.get().getY(gui.guiHeight());
 
         RenderSystem.setShader(GameRenderer::getPositionTexShader);
         RenderSystem.setShaderTexture(0, meterLoc);
@@ -79,28 +87,29 @@ public class MetalOverlay implements IGuiOverlay {
                 int i = mt.getIndex();
                 int offset = (i / 2) * 4; // Adding a gap between pairs
                 // Draw the bars first
-                blit(guiGraphics, renderX + 1 + (7 * i) + offset, renderY + 5 + metalY, 7 + (6 * i), 1 + metalY, 3, 10 - metalY);
+                blit(gui, renderX + 1 + (7 * i) + offset, renderY + 5 + metalY, 7 + (6 * i), 1 + metalY, 3,
+                     10 - metalY);
                 // Draw the gauges second, so that highlights and decorations show over the bar.
-                blit(guiGraphics, renderX + (7 * i) + offset, renderY, 0, 0, 5, 20);
+                blit(gui, renderX + (7 * i) + offset, renderY, 0, 0, 5, 20);
                 // Draw the fire if it is burning
                 if (data.isBurning(mt)) {
                     int frameCount = (currentFrame + i) % 4;
                     var frame = Frames[frameCount];
-                    blit(guiGraphics, renderX + (7 * i) + offset, renderY + 4 + metalY, frame.x, frame.y, 5, 3);
+                    blit(gui, renderX + (7 * i) + offset, renderY + 4 + metalY, frame.x, frame.y, 5, 3);
                 }
             }
 
         }
 
         // Update the animation counters
-
-        if (gui.getGuiTicks() % 3 == 0) {
+        if (mc.gui.getGuiTicks() % 6 == 0) {
             currentFrame++;
             if (currentFrame > 3) {
                 currentFrame = 0;
             }
         }
     }
+
 
     public enum SCREEN_LOC {
         TOP_RIGHT,
