@@ -2,8 +2,11 @@ package com.legobmw99.allomancy.modules.powers.network;
 
 import com.legobmw99.allomancy.Allomancy;
 import com.legobmw99.allomancy.modules.powers.data.AllomancerAttachment;
+import io.netty.buffer.ByteBuf;
+import net.minecraft.core.UUIDUtil;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.network.codec.ByteBufCodecs;
+import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
@@ -11,25 +14,21 @@ import net.minecraft.server.level.ServerPlayer;
 import java.util.UUID;
 
 public record AllomancerDataPayload(CompoundTag nbt, UUID player) implements CustomPacketPayload {
-    public static final ResourceLocation ID = new ResourceLocation(Allomancy.MODID, "player_data");
 
+    public static final Type<AllomancerDataPayload> TYPE =
+            new Type<>(ResourceLocation.fromNamespaceAndPath(Allomancy.MODID, "player_data"));
+
+    public static final StreamCodec<ByteBuf, AllomancerDataPayload> STREAM_CODEC =
+            StreamCodec.composite(ByteBufCodecs.COMPOUND_TAG, AllomancerDataPayload::nbt, UUIDUtil.STREAM_CODEC,
+                                  AllomancerDataPayload::player, AllomancerDataPayload::new);
 
     public AllomancerDataPayload(ServerPlayer player) {
-        this(player.getData(AllomancerAttachment.ALLOMANCY_DATA).serializeNBT(), player.getUUID());
-    }
-
-    public AllomancerDataPayload(final FriendlyByteBuf buf) {
-        this(buf.readNbt(), buf.readUUID());
+        this(player.getData(AllomancerAttachment.ALLOMANCY_DATA).serializeNBT(player.registryAccess()),
+             player.getUUID());
     }
 
     @Override
-    public void write(FriendlyByteBuf buf) {
-        buf.writeNbt(this.nbt);
-        buf.writeUUID(this.player);
-    }
-
-    @Override
-    public ResourceLocation id() {
-        return ID;
+    public Type<AllomancerDataPayload> type() {
+        return TYPE;
     }
 }
