@@ -5,21 +5,29 @@ import com.legobmw99.allomancy.modules.powers.client.network.ClientPayloadHandle
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.server.level.ServerPlayer;
 import net.neoforged.neoforge.network.PacketDistributor;
-import net.neoforged.neoforge.network.event.RegisterPayloadHandlerEvent;
-import net.neoforged.neoforge.network.registration.IPayloadRegistrar;
+import net.neoforged.neoforge.network.event.RegisterPayloadHandlersEvent;
+import net.neoforged.neoforge.network.handling.DirectionalPayloadHandler;
+import net.neoforged.neoforge.network.registration.PayloadRegistrar;
 
 
 public class Network {
-    public static void registerPayloads(final RegisterPayloadHandlerEvent event) {
-        final IPayloadRegistrar registrar = event.registrar(Allomancy.MODID).versioned("2.0");
+    public static void registerPayloads(final RegisterPayloadHandlersEvent event) {
+        final PayloadRegistrar registrar = event.registrar(Allomancy.MODID).versioned("3.0");
 
-        registrar.play(AllomancerDataPayload.ID, AllomancerDataPayload::new, handler -> handler.client(ClientPayloadHandler::updateAllomancer));
-        registrar.play(EmotionPayload.ID, EmotionPayload::new, handler -> handler.server(ServerPayloadHandler::changeEmotion));
-        registrar.play(BlockPushPullPayload.ID, BlockPushPullPayload::new, handler -> handler.server(ServerPayloadHandler::tryPushPullBlock));
-        registrar.play(EntityPushPullPayload.ID, EntityPushPullPayload::new, handler -> handler.server(ServerPayloadHandler::tryPushPullEntity));
-        registrar.play(ToggleBurnPayload.ID, ToggleBurnPayload::new, handler -> handler.server(ServerPayloadHandler::toggleBurnRequest));
-        registrar.play(EnhanceTimePayload.ID, EnhanceTimePayload::new,
-                       handler -> handler.server(ServerPayloadHandler::updateEnhanced).client(ClientPayloadHandler::updateEnhanced));
+        registrar.playToClient(AllomancerDataPayload.TYPE, AllomancerDataPayload.STREAM_CODEC,
+                               ClientPayloadHandler::updateAllomancer);
+        registrar.playToServer(EmotionPayload.TYPE, EmotionPayload.STREAM_CODEC, ServerPayloadHandler::changeEmotion);
+        registrar.playToServer(BlockPushPullPayload.TYPE, BlockPushPullPayload.STREAM_CODEC,
+                               ServerPayloadHandler::tryPushPullBlock);
+        registrar.playToServer(EntityPushPullPayload.TYPE, EntityPushPullPayload.STREAM_CODEC,
+                               ServerPayloadHandler::tryPushPullEntity);
+
+        registrar.playToServer(ToggleBurnPayload.TYPE, ToggleBurnPayload.STREAM_CODEC,
+                               ServerPayloadHandler::toggleBurnRequest);
+
+        registrar.playBidirectional(EnhanceTimePayload.TYPE, EnhanceTimePayload.STREAM_CODEC,
+                                    new DirectionalPayloadHandler<>(ClientPayloadHandler::updateEnhanced,
+                                                                    ServerPayloadHandler::updateEnhanced));
 
     }
 
@@ -28,6 +36,6 @@ public class Network {
     }
 
     public static void sync(CustomPacketPayload msg, ServerPlayer player) {
-        PacketDistributor.TRACKING_ENTITY_AND_SELF.with(player).send(msg);
+        PacketDistributor.sendToPlayersTrackingEntityAndSelf(player, msg);
     }
 }
